# springdoc-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/role/unassignPermission": {
      "post": {
        "tags": [
          "Role Controller"
        ],
        "summary": "Unassign permission from a role",
        "description": "Removes an existing permission from a role using the provided role and permission IDs.",
        "operationId": "unassignPermissionFromRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperRoleResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/role/create": {
      "post": {
        "tags": [
          "Role Controller"
        ],
        "summary": "Create a new role",
        "description": "Creates a new role with the specified name and description.",
        "operationId": "createRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperRoleResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/role/addPermission": {
      "post": {
        "tags": [
          "Role Controller"
        ],
        "summary": "Assign permission to a role",
        "description": "Adds a specific permission to the given role based on role and permission IDs.",
        "operationId": "addPermissionToRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWrapperRoleResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RolePermissionRequestDTO": {
        "required": [
          "permissionId",
          "roleId"
        ],
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseWrapperRoleResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RoleResponseDTO"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoleResponseDTO": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int32"
          },
          "roleName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateRoleDTO": {
        "required": [
          "roleName"
        ],
        "type": "object",
        "properties": {
          "roleName": {
            "pattern": "^[A-Z_]+$",
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    }
  }
}